backend/requirements.txt
Django==4.0.5
django_filter==21.1
djangorestframework==3.13.1
django-cors-headers==3.12.0
